requirements.txt
openai
streamlit 
PyPDF2 
python-docx 
pandas
python-dotenv

# Dependencies for PR Reviews
requests
tiktoken

# Dependencies for the 100 days of LLM course
torch